workforce_backend/requirements.txt
asgiref==3.8.1
bcrypt==4.3.0
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.16.0
PyJWT==2.10.1
sqlparse==0.5.3
typing_extensions==4.13.0
